daniilrusanov/am_i_wilson (change parameter from double to long)

+++ src/main/java/org/academy/kata/implementation/daniilrusanov/Eight.java
@@ -88,7 +88,7 @@ public class Eight implements IEight {
     }
 
     @Override
-    public boolean am_i_wilson(double n) {
+    public boolean am_i_wilson(long n) {
         if(!isPrime(n)) return false;
         BigInteger factorialVal = factorial(n-1);
         BigInteger modulus = BigInteger.valueOf(n).multiply(BigInteger.valueOf(n));
